# Conflicts:
#	classes/dsr/core/rtps/dsrparticipant.cpp
#	classes/dsr/python-wrapper/python_api.cpp

+++ classes/dsr/core/crdt/delta_crdt.h
@@ -16,7 +16,7 @@ Reinplementation from https://github.com/CBaquero/delta-enabled-crdts
 #include <type_traits>
 
 
-using key_type = uint32_t;
+using key_type = uint64_t;
 
 
 // Autonomous causal context, for context sharing in maps
